requirements_cpu.txt
ffmpeg-python==0.1.17
tqdm
tensorflow==1.13.1
imageio-ffmpeg

psutil
pathlib
numpy==1.17.0
opencv-python==4.1.0.25
scikit-image
Pillow==6.1.0
scikit-learn
toposort
fastcluster
matplotlib==2.2.2
imageio==2.5.0
ffmpy==0.2.2
# Revert back to nvidia-ml-py3 when windows/system32 patch is implemented
nvidia-ml-py3
#nvidia-ml-py3
h5py==2.9.0
Keras==2.2.4
# pywin32 ; sys_platform == "win32"
git+https://www.github.com/keras-team/keras-contrib.git